Consolidate scattered lsm files by ownership
Following the same pattern as the previous merge of picker/planner/pacer/
executor into level_manager, fold remaining single-topic files into
their owner files:

  landing.go           -> level_handler.go (levelHandler owns the buffer
                          and its RWMutex; landing is a thin adapter)
  shard_hint.go        -> shard.go (both are sharding concepts)
  l0_sublevels.go      -> level_handler.go (only levelHandler reads the
                          sublevel layout; lh.l0Sublevels field lives there)
  range_tombstone.go   -> split by receiver:
                          RangeTombstoneView + (lsm *LSM) methods   -> lsm.go
                          (lm *levelManager).rebuildRangeTombstones -> level_manager.go

Test files follow source 1:1:
  l0_sublevels_test.go     -> level_handler_test.go (rename)
  shard_hint_test.go       -> shard_test.go (merge)
  shard_hint_bench_test.go -> shard_bench_test.go (rename)

After this pass engine/lsm/ root has 17 source files (down from 21):

  compactdef.go        compaction.go        diagnostics.go
  errors.go            external_sst.go      flush_runtime.go
  iterator.go          level_handler.go     level_manager.go
  lsm.go               memtable.go          options.go
  shard.go             table.go             table_builder.go

with matching _test.go and (where present) _bench_test.go files.

All 60+ packages, architecture dependency check, and lsm subpackage
tests stay green.
