+++ engine/lsm/level_manager.go
@@ -1,6 +1,7 @@
 package lsm
 
 import (
+	"bytes"
 	"errors"
 	"fmt"
 	"log/slog"
@@ -1566,4 +1567,56 @@ func (lm *levelManager) subcompact(it index.Iterator, kr plan.KeyRange, cd compa
 			break
 		}
 	}
+}
+
+// rebuildRangeTombstones scans SST levels to repopulate the range tombstone
+// collector. Memtable tombstones are tracked separately in
+// memTable.rangeTombstones and must not be included here to avoid duplication
+// when those memtables flush. Called at startup and after max-level
+// compaction (which may drop tombstones).
+func (lm *levelManager) rebuildRangeTombstones() {
+	if lm == nil || lm.rtCollector == nil || len(lm.levels) == 0 {
+		return
+	}
+	var ranges []tombstone.Range
+	opt := &index.Options{IsAsc: true}
+	// Only scan SST levels — memtable tombstones are tracked separately
+	// in memTable.rangeTombstones and must not be duplicated here.
+	iters := lm.iterators(opt)
+	defer func() {
+		for _, it := range iters {
+			if it != nil {
+				_ = it.Close()
+			}
+		}
+	}()
+	for _, it := range iters {
+		if it == nil {
+			continue
+		}
+		it.Rewind()
+		for it.Valid() {
+			if item := it.Item(); item != nil {
+				if e := item.Entry(); e != nil && e.IsRangeDelete() {
+					cf, start, version, ok := kv.SplitInternalKey(e.Key)
+					if !ok {
+						it.Next()
+						continue
+					}
+					if bytes.Compare(start, e.RangeEnd()) >= 0 {
+						it.Next()
+						continue
+					}
+					ranges = append(ranges, tombstone.Range{
+						CF:      cf,
+						Start:   kv.SafeCopy(nil, start),
+						End:     kv.SafeCopy(nil, e.RangeEnd()),
+						Version: version,
+					})
+				}
+			}
+			it.Next()
+		}
+	}
+	lm.rtCollector.Rebuild(ranges)
 }
